Refactor: 문자열 계산기 패키지 이동 (lotto -> calculator)
