+++ src/main/java/calculator/domain/Expression.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package calculator.domain;
 
 import java.util.LinkedList;
 import java.util.List;
