+++ src/main/java/calculator/domain/ExpressionElement.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package calculator.domain;
 
 public abstract class ExpressionElement {
