+++ src/main/java/calculator/domain/MultiplyStrategy.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package calculator.domain;
 
 public class MultiplyStrategy implements OperationStrategy {
     @Override
