+++ src/main/java/calculator/domain/SubtractStrategy.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package calculator.domain;
 
 public class SubtractStrategy implements OperationStrategy {
     @Override
