+++ src/main/java/calculator/ui/ResultView.java
@@ -1,4 +1,4 @@
-package lotto.ui;
+package calculator.ui;
 
 import java.text.MessageFormat;
